test: updated test setup for e2e front end tests
