+++ client/e2e/protractor.conf.js
@@ -51,14 +51,6 @@ exports.config = {
     require('ts-node').register({
       project: require('path').join(__dirname, './tsconfig.e2e.json')
     });
-    var http = new HttpClient(browser.params.backendURL);
-    var testingUsersApi = `/testing/user/ivan-petrov-test@gmail.com`;
-    http.delete(testingUsersApi, {
-      Authorization: browser.params.testingCredentials
-    });
-    testingUsersApi = `/users/signin`;
-    const { testedUser } = require('@src-e2e/shared/tested-user');
-    http.post(testingUsersApi, testedUser);
 
     jasmine
       .getEnv()
