+++ client/e2e/src/auth/login.po.ts
@@ -38,10 +38,6 @@ export class LoginPage {
     submitForm();
   }
 
-  deleteTestedUser() {
-    this.registrationPage.deleteTestedUser();
-  }
-
   get registrationPageUrl(): string {
     return getPageUrl('registration');
   }
