+++ client/e2e/src/shared/index.ts
@@ -1,4 +1,5 @@
-export * from './functions.spec';
-export * from './tested-user';
 export * from './app-urls';
+export * from './functions.spec';
 export * from './reset-tested-user.spec';
+export * from './server-calls.spec';
+export * from './tested-user';
