+++ client/package.json
@@ -8,7 +8,8 @@
     "lint": "ng lint",
     "test:unit-test": "ng test --watch=false",
     "test:e2e": "ng e2e",
-    "test": "npm run test:unit-test && npm run test:e2e",
+    "server:test": "cd ../server && npm run start:test",
+    "e2e": "start-server-and-test server:test 3000 test:e2e",
     "start:dev": "ng serve --proxy-config proxy.conf.json",
     "postinstall": "ng build --aot --prod",
     "precommit": "lint-staged",
@@ -48,6 +49,7 @@
     "protractor-html-reporter-2": "^1.0.4",
     "protractor-http-client": "^1.0.4",
     "rxjs": "~6.3.3",
+    "start-server-and-test": "^1.11.0",
     "tslib": "^1.10.0",
     "typescript": "~3.2.2",
     "webdriver-manager": "^12.1.7",
