+++ client/src/environments/environment.staging.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
-  backendURL: 'https://family-expenses-staging.herokuapp.com/api/v1',
+  backendURL: 'http://localhost:3000/api/v1',
   supportEmail: 'orly.knop@gmail.com'
 };
