--- server/__tests__/e2e.ts
@@ -1,3 +1,0 @@
-test("Should make integration test of an app", () => {
-  expect(2).toBe(2);
-});
